add check

+++ apps/bot/src/bot/textCommands/player/join.ts
@@ -8,6 +8,7 @@ export default class Cmd extends TextCommand {
 	}
 
 	override async run(message: Message) {
+		if (message.channelId !== "1105754920063352852") return message.reply("LMAO get a load of this nerd")
 		await message.delete()
 		message.member?.roles.add(serverIds.roles.player)
 	}
